Ensures events can be type hinted
If an event lists argument types, any typed method being bound to this
event must have a matching signature, ignoring the return type. A fatal
error is thrown at the point of binding if not.

If either the event or the method are not type hinted, the bind can
take place without error and it is then up to the programmer to ensure
that no type errors occur.