Removes incorrect test regarding auto loading
Test existed that expected an error if an included script did not
contain the 'class' it was expecting. The problem being that we are not
exclusively loading classes and so errors were incorrectly triggered
when loading interfaces.

Perhaps a better solution can be found later but for now, the checking
has been removed.