Adds the ability to proxy a function in a method
This allows you to declare a function within a method call which acts
as though it is the method itself. This is specifically so that it can
access private or protected class members.

A likely use-case would be when passing a callback to some other code -
when the callback is run at a later point, you would like to still have
the same privileges as the original declaring method.