Enables optional method arguments
Method arguments can now be declared optional, indicated by either
appending a question mark to the argument type or by declaring a valid
default value (i.e. number = 30). The first method to be declared that
can handle the provided arguments will be called; if this is a method
with optional arguments, either null or the default value will be
supplied in place.