Fixes issue calling private method after require
Previously a request to require a file and then call a private class
method would fail as the actual method call would be made anonymously.
Now, the calling object is recorded on the require call and passed to
the member registry to handle the call. This means that it is subject to
the same access checks as a normal method call.