Removes redundant 'get' and 'set' methods
Originally, a property was manipulated using this.get('myProperty') and
this.set('myProperty', 'Value') however since the newer format (calling
a method of the same name) these have become redundant.

This removes the get and set methods so that these name become available
again for client methods, properties and events.