Allows abstract methods to omit return types
This applies to interface methods also (as they are abstract). Simply,
a method may be defined without the brackets containing argument types
and any method which matches in all other ways will be deemed a valid
implementation.

This has implications in dependency injection where one class may want
to specify a method name and return type but not care what values the
method needs; because it intends to inject the dependencies
dynamically.